+++ CHANGELOG.md
@@ -9,6 +9,11 @@ __Enhancements__
 __BugFixes__
 
 
+__Maintenance__
+
+- Removes old routing modes seppmail and microsoft. Now only 'inline' and 'parallel' is allowed.
+
+
 ## 1.1.0 ARC Signing update (22-08-2022)
 
 __Enhancements__